[lti] refactor lti to support multiple versions
