+++ inginious/frontend/flask/mapping.py
@@ -23,7 +23,8 @@ from inginious.frontend.pages.social import AuthenticationPage, CallbackPage
 from inginious.frontend.pages.course_register import CourseRegisterPage
 from inginious.frontend.pages.course import CoursePage
 from inginious.frontend.pages.tasks import TaskPage, TaskPageStaticDownload
-from inginious.frontend.pages.lti import LTITaskPage, LTILaunchPage, LTIBindPage, LTIAssetPage, LTILoginPage
+from inginious.frontend.pages.lti.v1_1 import LTI11LaunchPage, LTI11BindPage, LTI11LoginPage
+from inginious.frontend.pages.lti import LTITaskPage, LTIAssetPage
 from inginious.frontend.pages.group import GroupPage
 from inginious.frontend.pages.marketplace import MarketplacePage
 from inginious.frontend.pages.marketplace_course import MarketplaceCoursePage
@@ -85,9 +86,9 @@ def init_flask_mapping(flask_app):
     flask_app.add_url_rule('/preferences/profile', view_func=ProfilePage.as_view('profilepage'))
     flask_app.add_url_rule('/lti/task', view_func=LTITaskPage.as_view('ltitaskpage'))
     flask_app.add_url_rule('/lti/<courseid>/<taskid>',
-                           view_func=LTILaunchPage.as_view('ltilaunchpage'))
-    flask_app.add_url_rule('/lti/bind', view_func=LTIBindPage.as_view('ltibindpage'))
-    flask_app.add_url_rule('/lti/login', view_func=LTILoginPage.as_view('ltiloginpage'))
+                           view_func=LTI11LaunchPage.as_view('ltilaunchpage'))
+    flask_app.add_url_rule('/lti/bind', view_func=LTI11BindPage.as_view('ltibindpage'))
+    flask_app.add_url_rule('/lti/login', view_func=LTI11LoginPage.as_view('ltiloginpage'))
     flask_app.add_url_rule('/lti/asset/<path:asset_url>',
                            view_func=LTIAssetPage.as_view('ltiassetpage'))
     flask_app.add_url_rule('/admin/<courseid>',
