+++ inginious/frontend/pages/utils.py
@@ -27,7 +27,7 @@ from pymongo.database import Database
 
 from inginious.frontend.course_factory import CourseFactory
 from inginious.frontend.task_factory import TaskFactory
-from inginious.frontend.lti_outcome_manager import LTIOutcomeManager
+from inginious.frontend.lti.v1_1 import LTIOutcomeManager
 
 
 class INGIniousPage(MethodView):
